+++ CHANGELOG.md
@@ -106,7 +106,7 @@ The Koto project adheres to
     # 1006
     ```
 - Meta map improvements
-  - Arithmetic-assignment operators (`@+=`, `@*=`, etc.) can now be implemented 
+  - Compound assignment operators (`@+=`, `@*=`, etc.) can now be implemented 
     in meta maps and external values.
   - The function call operator (`@||`) can be implemented to values that behave 
     like functions.